+++ internal/provider/data_source_utils_yaml_merge.go
@@ -9,7 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"gopkg.in/yaml.v3"
 )
 
 var _ datasource.DataSource = (*yamlMergeDataSource)(nil)
@@ -89,7 +88,7 @@ func (d *yamlMergeDataSource) Read(ctx context.Context, req datasource.ReadReque
 		MergeMaps(data, merged, config.MergeListItems.ValueBool())
 	}
 
-	output, err := yaml.Marshal(merged)
+	output, err := YamlMarshal(merged)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error converting result to YAML",